# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "integrity.molt Security Scanner",
    "description": "AI-powered Solana security scanner. All reports Ed25519 cryptographically signed. Pay per scan via x402 protocol.",
    "version": "1.0.0",
    "contact": {"url": "https://t.me/intmolt_bot"}
  },
  "servers": [{"url": "https://intmolt.org/api/v2"}],
  "paths": {
    "/scan/quick": {
      "post": {
        "summary": "Quick on-chain security scan (0.50 USDC)",
        "description": "Scans any Solana address for basic risk assessment - account info, balance, owner program, Ed25519 signed report.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"address": {"type": "string", "description": "Solana address to scan"}}, "required": ["address"]}}}},
        "responses": {"200": {"description": "Scan report", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}, "report": {"type": "string"}, "signed": {"type": "object"}, "timestamp": {"type": "string"}}}}}}, "402": {"description": "Payment required"}}
      }
    },
    "/scan/deep": {
      "post": {
        "summary": "Comprehensive security audit (5.00 USDC)",
        "description": "Full security review with multi-agent swarm analysis - scanner, analyst, and reputation agents.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"address": {"type": "string"}}, "required": ["address"]}}}},
        "responses": {"200": {"description": "Deep audit report"}, "402": {"description": "Payment required"}}
      }
    },
    "/scan/token": {
      "post": {
        "summary": "Token launch audit (0.75 USDC)",
        "description": "Analyzes token mint - mint authority, freeze authority, top holder distribution, supply analysis, rug risk rating.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"address": {"type": "string", "description": "Token mint address"}}, "required": ["address"]}}}},
        "responses": {"200": {"description": "Token audit report"}, "402": {"description": "Payment required"}}
      }
    },
    "/scan/wallet": {
      "post": {
        "summary": "Wallet profiling (0.50 USDC)",
        "description": "Wallet age estimate, activity level, DeFi exposure, token holdings, risk classification.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"address": {"type": "string", "description": "Wallet address"}}, "required": ["address"]}}}},
        "responses": {"200": {"description": "Wallet profile report"}, "402": {"description": "Payment required"}}
      }
    },
    "/scan/pool": {
      "post": {
        "summary": "DeFi pool safety scan (0.50 USDC)",
        "description": "Liquidity depth, LP token distribution, Raydium/Orca/Meteora analysis, withdrawal risk.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"address": {"type": "string", "description": "Pool address"}}, "required": ["address"]}}}},
        "responses": {"200": {"description": "Pool safety report"}, "402": {"description": "Payment required"}}
      }
    },
    "/scan/evm-token": {
      "post": {
        "summary": "EVM token risk scan (0.75 USDC)",
        "description": "Scans an EVM token contract on Base, Ethereum, or Arbitrum. Performs honeypot detection, source code analysis (blacklist, fees, mint, pause, selfdestruct, proxy), ownership check, and contract age. Returns risk score 0-100 with Ed25519 signed report.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"address": {"type": "string", "description": "EVM contract address (0x…)"}, "chain": {"type": "string", "enum": ["base", "ethereum", "arbitrum"], "default": "base"}}, "required": ["address"]}}}},
        "responses": {
          "200": {"description": "EVM token scan report", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}, "chain": {"type": "string"}, "address": {"type": "string"}, "score": {"type": "integer", "minimum": 0, "maximum": 100}, "recommendation": {"type": "string"}, "findings": {"type": "array", "items": {"type": "object", "properties": {"label": {"type": "string"}, "severity": {"type": "string", "enum": ["critical","high","medium","low","info"]}, "category": {"type": "string"}}}}, "meta": {"type": "object"}, "report": {"type": "string"}, "signed": {"type": "object"}, "timestamp": {"type": "string"}}}}}},
          "402": {"description": "Payment required (0.75 USDC via x402)"}
        }
      }
    },
    "/scan/agent-token": {
      "post": {
        "summary": "Agent Token Security Scan (0.15 USDC) — NEW",
        "description": "Scans a Metaplex Core Agent Token. Checks: Core NFT program ownership, treasury PDA, update authority risk (single key vs DAO/multisig), creator royalties, collection governance, and activity age. Returns risk score 0-100 with Ed25519 signed report. Metaplex Agent Tokens launched 2026-04-13.",
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"mint": {"type": "string", "description": "Metaplex Core asset address (base58)"}}, "required": ["mint"]}}}},
        "responses": {
          "200": {"description": "Agent token scan report", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}, "scan_type": {"type": "string"}, "target": {"type": "string"}, "score": {"type": "integer", "minimum": 0, "maximum": 100}, "risk_level": {"type": "string", "enum": ["LOW","MEDIUM","HIGH","CRITICAL"]}, "summary": {"type": "string"}, "findings": {"type": "array"}, "agent_metadata": {"type": "object"}, "token_metrics": {"type": "object"}, "signed": {"type": "object"}, "timestamp": {"type": "string"}}}}}},
          "402": {"description": "Payment required (0.15 USDC via x402)"}
        }
      }
    },
    "/services": {
      "get": {
        "summary": "Service discovery (free)",
        "description": "Lists all available services, pricing, and payment details.",
        "responses": {"200": {"description": "Service list"}}
      }
    },
    "/health": {
      "get": {
        "summary": "Health check (free)",
        "description": "Returns service status.",
        "responses": {"200": {"description": "Status OK"}}
      }
    }
  }
}